feat: implement esm.mjs
